api/requirements.txt
fastapi>=0.105.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0
redis>=4.5.0
pydantic>=2.0.0
alembic>=1.12.0
uvicorn>=0.23.0
python-jose>=3.4.0
argon2-cffi>=23.1.0
python-multipart>=0.0.20
sqlmodel>=0.0.24
passlib>=1.7.4
bcrypt>=4.3.0
celery>=5.3.0
# Type checking
mypy>=1.7.0
types-python-jose>=3.3.0
types-passlib>=1.7.7
sqlalchemy-stubs>=0.4.0